pyproject.toml
[project]
name = "devops-assistant"
version = "0.1.0"
description = "AI-powered DevOps Teaching Assistant"
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "pydantic>=1.8.2",
    "supabase>=0.7.1",
    "python-dotenv>=0.19.0",
    "pytest>=6.2.5",
    "httpx>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --cov=src"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3